+++ src/Concordium/Client/Output.hs
@@ -1094,6 +1094,7 @@ showRejectReason verbose = \case
     Types.InsufficientDelegationStake -> "not allowed to add delegator with 0 stake"
     Types.DelegationTargetNotABaker bid -> printf "delegation target %s is not a validator id" (show bid)
     Types.NonExistentTokenId tokenId -> printf "token id %s does not exist on-chain" (show tokenId)
+    Types.TokenHolderTransactionFailed reason -> printf "the token holder transaction failed with reason %s" (show reason)
 
 -- CONSENSUS