+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add `raw GetTokenList` command to display the list of protocol level tokens.
 - Subcommand `raw GetNextUpdateSequenceNumbers` now displays the next update sequence number for protocol level tokens chain updates.
 - The `consensus chain-update` command supports creating protocol level tokens now.